moving schema and relation types into shared package
