+++ packages/shared/src/types/relations.ts
@@ -1,18 +1,27 @@
+import { ForeignKey } from 'knex-schema-inspector/dist/types/foreign-key';
+
+export type RelationMeta = {
+	id: number;
+
+	many_collection: string;
+	many_field: string;
+
+	one_collection: string | null;
+	one_field: string | null;
+	one_collection_field: string | null;
+	one_allowed_collections: string[] | null;
+	one_deselect_action: 'nullify' | 'delete';
+
+	junction_field: string | null;
+	sort_field: string | null;
+
+	system?: boolean;
+};
+
 export type Relation = {
 	collection: string;
 	field: string;
 	related_collection: string | null;
-	meta?: {
-		id: number;
-		many_collection: string;
-		many_field: string;
-		one_collection: string;
-		one_field: null | string;
-		junction_field: null | string;
-		sort_field: null | string;
-		one_deselect_action: 'nullify' | 'delete';
-		one_collection_field: null | string;
-		one_allowed_collections: null | string[];
-	} | null;
-	schema?: Record<string, unknown> | null;
+	schema: ForeignKey | null;
+	meta: RelationMeta | null;
 };
